extract sortCandidates hook
